renamed ItemChecker class to ItemFilter and revised documentation
